bug #1992
fixed part of the bug related to status bar running .
now when exception is thrown status bar stoped and MessageDialog will be desplayed with message " MINDY obtained no results.", "see errors, logs" also added check if Mindy results are null, same action.

Better solution will be for aracne.jar to pass error message in this case it will be "Cannot find the hub probe: " + p.hub + ", nothing to be computed!" instead of or in addition to printing it to System.err, then mindy.jar will pass it to mindy plugin and mindy plugin will display it to user. As for now user will need to go to log or console to find out the reason why there are no results.


git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@5543 8b29bb6e-589a-435a-9a12-3ab907be3462